[refactor] Api 와 DataSource 분리
