--- app/src/main/java/com/mashup/zuzu/data/source/remote/WorldCupRemoteDataSource.kt
@@ -1,6 +1,0 @@
-package com.mashup.zuzu.data.source.remote
-
-/**
- * @Created by 김현국 2022/08/08
- */
-interface WorldCupRemoteDataSource
