+++ app/src/main/java/com/mashup/zuzu/data/source/remote/wine/WineApi.kt
@@ -1,4 +1,4 @@
-package com.mashup.zuzu.data.source.remote
+package com.mashup.zuzu.data.source.remote.wine
 
 import com.mashup.zuzu.data.response.GetDrinksResponse
 import com.mashup.zuzu.data.response.GetDrinksWithIdResponse
@@ -9,7 +9,7 @@ import retrofit2.http.Path
 /**
  * @Created by 김현국 2022/08/08
  */
-interface WineRemoteDataSource {
+interface WineApi {
 
     // 전체 주류 가져오기
     @GET("/drinks")
@@ -17,6 +17,6 @@ interface WineRemoteDataSource {
 
     // 특정 id에 해당하는 주류 디테일 정보 가져오기
     @GET("/drinks/{id}")
-    suspend fun getGetDrinksWithId(@Path("id") id: Int):
+    suspend fun getDrinksWithId(@Path("id") id: Int):
         Response<GetDrinksWithIdResponse>
 }
