+++ app/src/main/java/com/mashup/zuzu/data/source/remote/worldcup/WorldCupApi.kt
@@ -1,0 +1,6 @@
+package com.mashup.zuzu.data.source.remote.worldcup
+
+/**
+ * @Created by 김현국 2022/08/08
+ */
+interface WorldCupApi
